Move time_map.py out of enkf/util
