--- src/ert/_c_wrappers/enkf/util/__init__.py
@@ -1,3 +1,0 @@
-from .time_map import TimeMap
-
-__all__ = ["TimeMap"]
